# emdad-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Emdad Platform API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1_0/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "get All Products",
                "description": "get All Products Here",
                "operationId": "getAllProducts",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "": {
                                        "type": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "MainProduct": {
                                            "type": "integer",
                                            "example": "{'id': 2, 'name': 'LG','salary': 10000, 'parent_id': 1,'company_id': 1}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "update Product",
                "description": "update Product Here",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated Successfully"
                    },
                    "500": {
                        "description": "error"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "create Product",
                "description": "create Product Here",
                "operationId": "createProduct",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "categoryId",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "categoryId": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "created Successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "500": {
                        "description": "system error"
                    }
                }
            }
        },
        "/api/v1_0/products/get-By-Id/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "get By ProductId",
                "description": "get By ProductId Here",
                "operationId": "getByProductId",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ProductById": {
                                            "type": "integer",
                                            "example": "{'id': 2, 'name': 'LG','salary': 10000, 'parent_id': 1,'company_id': 1}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1_0/products/delete/{id}": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "delete Product",
                "description": "delete Product Here",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "301": {
                        "description": "deleted successfully"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1_0/products/restore/{id}": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "restore By ProductId",
                "description": "restore By ProductId Here",
                "operationId": "restoreByProductId",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "restored successfully"
                    },
                    "500": {
                        "description": "Resource Not Found"
                    },
                    "404": {
                        "description": "system error"
                    }
                }
            }
        },
        "/api/v1_0/products/company-products": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "set company product",
                "description": "Set Product or products list to a company profile",
                "operationId": "companyproduct",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "productList"
                                ],
                                "properties": {
                                    "productList": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "restored successfully"
                    },
                    "500": {
                        "description": "Resource Not Found"
                    },
                    "404": {
                        "description": "system error"
                    }
                }
            }
        },
        "/api/v1_0/products/change-Product-Status": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "set change Product Status",
                "description": "set change Product Status Here",
                "operationId": "changeProductStatus",
                "parameters": [
                    {
                        "name": "x-authorization",
                        "in": "header",
                        "description": "Set x-authorization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Set user authentication token",
                        "schema": {
                            "type": "beraer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "restored successfully"
                    },
                    "500": {
                        "description": "Resource Not Found"
                    },
                    "404": {
                        "description": "system error"
                    }
                }
            }
        },
...
    },
...
}